requirements.txt
pytest
requests
pytest-xdist
jsonschema
python-dotenv
pytest-html
allure-pytest
schemathesis